refactor(frontend): split monolithic types/index.tsx into domain-specific modules
Split 598-line types/index.tsx into:
- types/todo.ts: Todo, Tag, TodoItem, TodoTemplate
- types/execution.tsx: LogEntry, ExecutionRecord, ExecutorOption, EXECUTORS
- types/config.tsx: Config, ExecutorConfig, SlashCommandRule
- types/stats.ts: DashboardStats, SkillsStats, BackupStats, UsageStats

Named .tsx only for files containing JSX (execution.tsx, config.tsx).
Backward compatibility maintained via re-exports in index.tsx.
